Avoid integer narrowing in contraction hierarchies

+++ core/contraction.h
@@ -14,13 +14,13 @@ struct CH {
 
 CHPath* chpNew( int n, long length ) ;
 
-int chpLength( CHPath* this ) ;
+long chpLength( CHPath* this ) ;
 
 CHPath* chpCombine( CHPath* a, CHPath* b ) ;
 
 void chpDestroy( CHPath* this ) ;
     
-CHPath* dist( Graph *gg, char* from_v_label, char* to_v_label, WalkOptions *wo, int weightlimit, int return_full_path ) ;
+CHPath* dist( Graph *gg, char* from_v_label, char* to_v_label, WalkOptions *wo, long weightlimit, int return_full_path ) ;
 
 CHPath** get_shortcuts( Graph *gg, Vertex* vv, WalkOptions* wo, int search_limit, int* n ) ;
